Add a box definition object and an entity class
The spike revealed that although the gross parsing part is pretty
straightforward, to deal with nested boxes is much more complex,
particularly because there are so many box types with many different
rules for what attributes go where and how to unpack the data.

This box definition class and the Entity::Common module are a way of
making it easy to generate all the entity classes I'll need, as well as
giving me a hook to hang a parallel set of deserialization functions on.