Implement enough parser to parse a single box
Add basic Parser::Box instance that can parse the boxtype and size
fields from a box and hand off to a boxtype-specific deserializer for
the rest.

Automatically add boxtype and size attributes to all box Entities.

Allow Entities to be tested for equality, by a simple comparison of
attributes. (NB: I originally had comparison of the class of the entity
in `#==()` method body, but removed it since I currently have no other
real entities so there's no simple failing test to write that checks
that...)

Use the Structure::Definition class to define a working implementation
of the parser and deserializer for the `ftyp` box.

Remove all traces of the original parsing spike. (We're not yet at
parse-all-top-level-boxes feature parity, but we're parsing the
internals of boxes now.)