add all the child boxes to meta
Fill out the child boxes for meta.

Encounter the problem of XML boxes, whose boxtype 'xml' is not 4
characters. Obviously, it's a 32-bit value we're supposed to pretend is
a string, and the last byte is 32, ASCII space. So, the boxtype is
really 'xml '. Have to deal with that next...